refactor: DFIBalanceCard from BalanceScreen
