+++ mobile-app/app/screens/AppNavigator/screens/Balances/BalancesScreen.tsx
@@ -1,4 +1,3 @@
-import { IconButton } from '@components/IconButton'
 import { getNativeIcon } from '@components/icons/assets'
 import { View } from '@components/index'
 import {
@@ -20,16 +19,14 @@ import { ocean } from '@store/ocean'
 import { DFITokenSelector, DFIUtxoSelector, WalletToken } from '@store/wallet'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
-import DFIBackground from '@assets/images/DFI_balance_background.png'
-import DFIBackgroundDark from '@assets/images/DFI_balance_background_dark.png'
 import BigNumber from 'bignumber.js'
 import * as React from 'react'
 import { useCallback, useEffect, useState } from 'react'
-import { ImageBackground, RefreshControl, TouchableOpacity } from 'react-native'
+import { RefreshControl } from 'react-native'
 import NumberFormat from 'react-number-format'
 import { useDispatch, useSelector } from 'react-redux'
 import { BalanceParamList } from './BalancesNavigator'
-import { useThemeContext } from '@contexts/ThemeProvider'
+import { DFIBalanceCard } from './components/DFIBalanceCard'
 
 type Props = StackScreenProps<BalanceParamList, 'BalancesScreen'>
 
@@ -40,7 +37,6 @@ export function BalancesScreen ({ navigation }: Props): JSX.Element {
   const [refreshing, setRefreshing] = useState(false)
   const dispatch = useDispatch()
   const { wallets } = useWalletPersistenceContext()
-  const { isLight } = useThemeContext()
 
   useEffect(() => {
     dispatch(ocean.actions.setHeight(height))
@@ -55,23 +51,6 @@ export function BalancesScreen ({ navigation }: Props): JSX.Element {
   const tokens = useTokensAPI()
   const DFIToken = useSelector((state: RootState) => DFITokenSelector(state.wallet))
   const DFIUtxo = useSelector((state: RootState) => DFIUtxoSelector(state.wallet))
-  const onConvertPress = (): void => {
-    navigation.navigate({
-      name: 'Convert',
-      params: { mode: 'utxosToAccount' },
-      merge: true
-    })
-  }
-  const onSendPress = (): void => {
-    navigation.navigate({
-      name: 'Send',
-      params: { token: DFIUtxo },
-      merge: true
-    })
-  }
-  const onUtxoVsTokenPress = (): void => {
-    navigation.navigate('TokensVsUtxo')
-  }
 
   return (
     <ThemedFlatList
@@ -87,10 +66,6 @@ export function BalancesScreen ({ navigation }: Props): JSX.Element {
           <DFIBalanceCard
             utxo={DFIUtxo}
             token={DFIToken}
-            isLight={isLight}
-            onConvertPress={onConvertPress}
-            onSendPress={onSendPress}
-            onUtxoVsTokenPress={onUtxoVsTokenPress}
           />
           <ThemedSectionTitle
             testID='balances_title'
@@ -191,128 +166,5 @@ function BalanceItemRow ({
         </View>
       </View>
     </ThemedTouchableOpacity>
-  )
-}
-
-interface DFIBalanceCardProps {
-  utxo: WalletToken
-  token: WalletToken
-  onConvertPress: () => void
-  onSendPress: () => void
-  onUtxoVsTokenPress: () => void
-  isLight: boolean
-}
-
-function DFIBalanceCard (props: DFIBalanceCardProps): JSX.Element {
-  const DFIIcon = getNativeIcon('_UTXO')
-  const totalDFI = new BigNumber(props.utxo.amount).plus(new BigNumber(props.token.amount)).toFixed(8)
-
-  return (
-    <ThemedView
-      light={tailwind('bg-white border-gray-100')}
-      dark={tailwind('bg-gray-800')}
-      style={tailwind('mx-2 mt-4 rounded-lg flex-1')}
-      testID='dfi_balance_card'
-    >
-      <ImageBackground
-        source={props.isLight ? DFIBackground : DFIBackgroundDark}
-        style={tailwind('flex-1 rounded-lg overflow-hidden')}
-        resizeMode='cover'
-        resizeMethod='scale'
-      >
-        <View style={tailwind('flex-col flex-1 mx-4 mt-5 mb-4')}>
-          <View style={tailwind('flex-row pb-3 items-center')}>
-            <DFIIcon width={24} height={24} style={tailwind('mr-2')} />
-            <ThemedText style={tailwind('pr-9 text-lg font-bold')} testID='total_dfi_label'>DFI</ThemedText>
-            <ThemedText testID='total_dfi_amount'>{totalDFI} DFI</ThemedText>
-          </View>
-
-          <View style={tailwind('flex-row pb-1.5')}>
-            <ThemedText
-              light={tailwind('text-gray-500')}
-              dark={tailwind('text-gray-400')}
-              style={tailwind('pr-16 text-sm')}
-              testID='dfi_utxo_label'
-            >
-              UTXO
-            </ThemedText>
-            <ThemedText
-              light={tailwind('text-gray-500')}
-              dark={tailwind('text-gray-400')}
-              style={tailwind('text-sm')}
-              testID='dfi_utxo_amount'
-            >
-              {new BigNumber(props.utxo.amount).toFixed(8)}
-            </ThemedText>
-          </View>
-
-          <View style={tailwind('flex-row items-center flex-1')}>
-            <ThemedText
-              light={tailwind('text-gray-500')}
-              dark={tailwind('text-gray-400')}
-              style={tailwind('pr-14 text-sm')}
-              testID='dfi_token_label'
-            >
-              Token
-            </ThemedText>
-            <ThemedText
-              light={tailwind('text-gray-500')}
-              dark={tailwind('text-gray-400')}
-              style={tailwind('pl-1.5 text-sm')}
-              testID='dfi_token_amount'
-            >
-              {new BigNumber(props.token.amount).toFixed(8)}
-            </ThemedText>
-          </View>
-
-          <View style={tailwind('flex-row')}>
-            <UtxoVsTokensInfo onPress={props.onUtxoVsTokenPress} />
-            <View style={tailwind('flex-row flex-grow justify-end')}>
-              <IconButton
-                iconName='swap-vert'
-                iconSize={24}
-                iconType='MaterialIcons'
-                onPress={props.onConvertPress}
-                testID='convert_dfi_button'
-                style={tailwind('mr-2')}
-              />
-              <IconButton
-                iconName='arrow-upward'
-                iconSize={24}
-                iconType='MaterialIcons'
-                onPress={props.onSendPress}
-                testID='send_dfi_button'
-              />
-            </View>
-          </View>
-        </View>
-      </ImageBackground>
-    </ThemedView>
-  )
-}
-
-function UtxoVsTokensInfo (props: {onPress: () => void}): JSX.Element {
-  return (
-    <TouchableOpacity
-      onPress={props.onPress}
-      style={tailwind('flex-row items-end justify-start w-9/12')}
-      testID='token_vs_utxo_info'
-    >
-      <ThemedIcon
-        dark={tailwind('text-darkprimary-500')}
-        iconType='MaterialIcons'
-        light={tailwind('text-primary-500')}
-        name='help'
-        size={16}
-      />
-
-      <ThemedText
-        dark={tailwind('text-darkprimary-500')}
-        light={tailwind('text-primary-500')}
-        style={tailwind('ml-1 text-xs font-medium px-1')}
-      >
-        {translate('screens/ConvertScreen', 'Learn more about UTXO and tokens')}
-      </ThemedText>
-    </TouchableOpacity>
   )
 }
